requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
py-cpuinfo==9.0.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
requests==2.32.3
shiboken6==6.7.2
sseclient-py==1.8.0
urllib3==2.2.2
